feat: move isSameJsonMock guard to core package
