--- packages/msw-devtools/src/utils/isSameJsonMock.ts
@@ -1,5 +1,0 @@
-import type { JsonMock } from "core"
-
-export const isSameJsonMock = (a: JsonMock, b: JsonMock) => {
-  return a.url === b.url && a.method === b.method
-}
